refactor: split Player.cpp into multiple logic files
