+++ GameMoi/GameMoi.vcxproj
@@ -134,7 +134,7 @@
     <ClInclude Include="Camera\Camera.h" />
     <ClInclude Include="Models\Characters\Enemy.h" />
     <ClInclude Include="Models\Characters\Info.h" />
-    <ClInclude Include="Models\Characters\Player.h" />
+    <ClInclude Include="Models\Characters\Player\Player.h" />
     <ClInclude Include="Models\GameObject.h" />
     <ClInclude Include="Models\Weapons\Axe.h" />
     <ClInclude Include="Models\Weapons\Weapon.h" />
@@ -159,7 +159,11 @@
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Models\Characters\Enemy.cpp" />
     <ClCompile Include="Models\Characters\Info.cpp" />
-    <ClCompile Include="Models\Characters\Player.cpp" />
+    <ClCompile Include="Models\Characters\Player\Player.cpp" />
+    <ClCompile Include="Models\Characters\Player\PlayerCombat.cpp" />
+    <ClCompile Include="Models\Characters\Player\PlayerMove.cpp" />
+    <ClCompile Include="Models\Characters\Player\PlayerStair.cpp" />
+    <ClCompile Include="Models\Characters\Player\PlayerUpdate.cpp" />
     <ClCompile Include="Models\GameObject.cpp" />
     <ClCompile Include="Models\Weapons\Axe.cpp" />
     <ClCompile Include="Models\Weapons\Weapon.cpp" />
